Revises Timestamp offset fields to match data model
For year/month/day, the field has been removed, since it can always be assumed to be unknown.
For minute/second/fractional_second it is now Option<i16> to reflect that it could be unknown (None), and only needs to store at most the number of minutes in a day. Option is necessary because Rust does not distinguish between positive and negative ints.

See also https://github.com/amzn/ion-docs/issues/151.